solutions oop-classes-and-objects task rectangle

+++ src/com/rakovets/course/challenge/classes/Point.java
@@ -4,7 +4,7 @@ public class Point {
     double x;
     double y;
 
-    Point(int x, int y) {
+    Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
